rename usb_util_windows to usb_util
