+++ source/main.py
@@ -7,7 +7,7 @@ import json
 import os
 import sys
 
-import usb_util_windows
+import usb_util
 
 class SidebarFrame(ctk.CTkFrame):
     def __init__(self, master, app_instance, **kwargs):#Widget Placement
@@ -218,7 +218,7 @@ class App(ctk.CTk):
     
     def get_device_name(self, serial_number):#Function to determine device name from serial
         try:
-            hmds = list(usb_util_windows.find_hmd())
+            hmds = list(usb_util.find_hmd())
             for hmd in hmds:
                 dongle_serials = hmd.get_dongle_serials()
                 if serial_number in dongle_serials:
